Rename file to match name of hook inside it
